+++ app/(server)/pubhub/v1/products/[identifier]/(lib)/requests.ts
@@ -1,4 +1,4 @@
-import { TUserInfo } from "@/app/pubhub/(lib)/types"
+import { TUserInfo } from "@/app/(server)/pubhub/(lib)/types"
 import { getPublizonServiceParameters } from "@/lib/helpers/publizon"
 import { createClientAsync } from "@/lib/soap/publizon/v2_7/generated/getproduct"
