+++ app/(server)/pubhub/v1/products/[identifier]/(lib)/schemas.ts
@@ -1,6 +1,6 @@
 import { z } from "zod"
 
-import { regexDate } from "@/app/pubhub/helper"
+import { regexDate } from "@/app/(server)/pubhub/helper"
 
 export const getProductsSchema = z.object({
   response: z.object({
